# test.github.io-master/server/src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 云上共富 - 农村集体经济数字化平台 API
  description: |
    云上共富购物系统后端API文档

    ## 核心功能模块
    - 用户认证模块 (/api/auth)
    - 商品管理模块 (/api/products)
    - 订单模块 (/api/orders)
    - 投资项目模块 (/api/projects)
    - 投标模块 (/api/bids)
    - 用户管理模块 (/api/users)
  version: 1.0.0
  contact:
    name: 云上共富技术支持
servers:
  - url: http://localhost:3001
    description: 开发环境服务器
...
paths:
...
  /api/projects:
    get:
      tags:
        - 项目
      summary: 获取投资项目列表
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: type
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [交易中, 已完成, 已取消]
        - name: keyword
          in: query
          schema:
            type: string
      responses:
        '200':
          description: 成功
    post:
      tags:
        - 项目
      summary: 创建投资项目（仅商户）
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - price
              properties:
                name:
                  type: string
                description:
                  type: string
                price:
                  type: number
                location:
                  type: string
                area:
                  type: string
                type:
                  type: string
                contactPerson:
                  type: string
                contactInfo:
                  type: string
                bidEndDate:
                  type: string
                  format: date-time
      responses:
        '201':
          description: 创建成功
  /api/projects/{id}:
    get:
      tags:
        - 项目
      summary: 获取项目详情
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 成功
    put:
      tags:
        - 项目
      summary: 更新项目（仅所有者）
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 更新成功
    delete:
      tags:
        - 项目
      summary: 删除项目（仅所有者）
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 删除成功
...
  /api/users:
    get:
      tags:
        - 用户
      summary: 获取用户列表（仅管理员）
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 成功
  /api/users/{id}:
    get:
      tags:
        - 用户
      summary: 获取用户详情
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 成功
  /api/users/profile:
    put:
      tags:
        - 用户
      summary: 更新个人信息
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                phone:
                  type: string
                address:
                  type: string
                businessLicense:
                  type: string
                businessScope:
                  type: string
                companyPhone:
                  type: string
                companyAddress:
                  type: string
      responses:
        '200':
          description: 更新成功
  /api/users/password:
    put:
      tags:
        - 用户
      summary: 修改密码
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - currentPassword
                - newPassword
                - confirmPassword
              properties:
                currentPassword:
                  type: string
                newPassword:
                  type: string
                confirmPassword:
                  type: string
      responses:
        '200':
          description: 密码修改成功
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...